saveHoadon: check nhanVienThucHien!=null

+++ beautify_api/8.0.0/aspnet-core/src/BanHangBeautify.Application/HoaDon/HoaDon/HoaDonAppService.cs
@@ -80,21 +80,23 @@ namespace BanHangBeautify.HoaDon.HoaDon
                     ctNew.CreationTime = DateTime.Now;
                     lstCTHoaDon.Add(ctNew);
 
-                    foreach (var nvth in item.nhanVienThucHien)
+                    if (item.nhanVienThucHien != null)
                     {
-                        BH_NhanVienThucHien nvNew = ObjectMapper.Map<BH_NhanVienThucHien>(nvth);
-                        nvNew.Id = Guid.NewGuid();
-                        nvNew.IdHoaDonChiTiet = ctNew.Id;
-                        nvNew.TenantId = AbpSession.TenantId ?? 1;
-                        nvNew.CreatorUserId = AbpSession.UserId;
-                        nvNew.CreationTime = DateTime.Now;
-                        lstNVTH.Add(nvNew);
+                        foreach (var nvth in item.nhanVienThucHien)
+                        {
+                            BH_NhanVienThucHien nvNew = ObjectMapper.Map<BH_NhanVienThucHien>(nvth);
+                            nvNew.Id = Guid.NewGuid();
+                            nvNew.IdHoaDonChiTiet = ctNew.Id;
+                            nvNew.TenantId = AbpSession.TenantId ?? 1;
+                            nvNew.CreatorUserId = AbpSession.UserId;
+                            nvNew.CreationTime = DateTime.Now;
+                            lstNVTH.Add(nvNew);
+                        }
                     }
-                    
                 }
                 await _hoaDonRepository.InsertAsync(objHD);
                 await _hoaDonChiTietRepository.InsertRangeAsync(lstCTHoaDon);
-                if (lstNVTH!=null && lstNVTH.Count > 0)
+                if (lstNVTH != null && lstNVTH.Count > 0)
                 {
                     await _nvThucHien.InsertRangeAsync(lstNVTH);
                 }
@@ -168,17 +170,20 @@ namespace BanHangBeautify.HoaDon.HoaDon
                     ctUp.CreationTime = DateTime.Now;
                     await _hoaDonChiTietRepository.UpdateAsync(ctUp);
 
-                    _nvthService.DeleteNVThucHienDichVu(item.Id);
-
-                    foreach (var nvth in item.nhanVienThucHien)
+                    if (item.nhanVienThucHien != null)
                     {
-                        BH_NhanVienThucHien nvNew = ObjectMapper.Map<BH_NhanVienThucHien>(nvth);
-                        nvNew.Id = Guid.NewGuid();
-                        nvNew.IdHoaDonChiTiet = ctUp.Id;
-                        nvNew.TenantId = AbpSession.TenantId ?? 1;
-                        nvNew.CreatorUserId = AbpSession.UserId;
-                        nvNew.CreationTime = DateTime.Now;
-                        lstNVTH.Add(nvNew);
+                        _nvthService.DeleteNVThucHienDichVu(item.Id);
+
+                        foreach (var nvth in item.nhanVienThucHien)
+                        {
+                            BH_NhanVienThucHien nvNew = ObjectMapper.Map<BH_NhanVienThucHien>(nvth);
+                            nvNew.Id = Guid.NewGuid();
+                            nvNew.IdHoaDonChiTiet = ctUp.Id;
+                            nvNew.TenantId = AbpSession.TenantId ?? 1;
+                            nvNew.CreatorUserId = AbpSession.UserId;
+                            nvNew.CreationTime = DateTime.Now;
+                            lstNVTH.Add(nvNew);
+                        }
                     }
                 }
                 await _hoaDonRepository.UpdateAsync(objOld);
@@ -201,88 +206,143 @@ namespace BanHangBeautify.HoaDon.HoaDon
         public async Task<CreateHoaDonDto> Update_InforHoaDon(CreateHoaDonDto objUp)
         {
             BH_HoaDon objOld = await _hoaDonRepository.FirstOrDefaultAsync(objUp.Id);
-            if (string.IsNullOrEmpty(objUp.MaHoaDon))
+            if (objOld != null)
             {
-                objOld.MaHoaDon = await _repoHoaDon.GetMaHoaDon(AbpSession.TenantId ?? 1, objUp.IdChiNhanh, objUp.IdLoaiChungTu, objUp.NgayLapHoaDon);
+                if (string.IsNullOrEmpty(objUp.MaHoaDon))
+                {
+                    objOld.MaHoaDon = await _repoHoaDon.GetMaHoaDon(AbpSession.TenantId ?? 1, objUp.IdChiNhanh, objUp.IdLoaiChungTu, objUp.NgayLapHoaDon);
+                }
+                else
+                {
+                    objOld.MaHoaDon = objUp.MaHoaDon;
+                }
+                objOld.NgayLapHoaDon = objUp.NgayLapHoaDon;
+                objOld.NgayApDung = objUp.NgayApDung;
+                objOld.NgayHetHan = objUp.NgayHetHan;
+                objOld.IdChiNhanh = objUp.IdChiNhanh;
+                objOld.IdKhachHang = objUp.IdKhachHang;
+                objOld.IdNhanVien = objUp.IdNhanVien;
+                objOld.IdPhong = objUp.IdViTriPhong;
+                objOld.IdHoaDon = objUp.IdHoaDon;
+                objOld.TongTienHangChuaChietKhau = objUp.TongTienHangChuaChietKhau;
+                objOld.PTChietKhauHang = objUp.PTChietKhauHang;
+                objOld.TongChietKhauHangHoa = objUp.TongChietKhauHangHoa;
+                objOld.TongTienHang = objUp.TongTienHang;
+                objOld.PTThueHD = objUp.PTThueHD;
+                objOld.TongTienThue = objUp.TongTienThue;
+                objOld.TongTienHDSauVAT = objUp.TongTienHDSauVAT;
+                objOld.PTGiamGiaHD = objUp.PTGiamGiaHD;
+                objOld.TongGiamGiaHD = objUp.TongGiamGiaHD;
+                objOld.ChiPhiTraHang = objUp.ChiPhiTraHang;
+                objOld.TongThanhToan = objUp.TongThanhToan;
+                objOld.ChiPhiHD = objUp.ChiPhiHD;
+                objOld.ChiPhi_GhiChu = objUp.ChiPhi_GhiChu;
+                objOld.DiemGiaoDich = objUp.DiemGiaoDich;
+                objOld.GhiChuHD = objUp.GhiChuHD;
+                objOld.TrangThai = objUp.TrangThai;
+                objOld.LastModifierUserId = AbpSession.UserId;
+                objOld.LastModificationTime = DateTime.Now;
+
+                await _hoaDonRepository.UpdateAsync(objOld);
+                var dataHD = ObjectMapper.Map<CreateHoaDonDto>(objOld);
+                return dataHD;
             }
-            else
-            {
-                objOld.MaHoaDon = objUp.MaHoaDon;
-            }
-            objOld.NgayLapHoaDon = objUp.NgayLapHoaDon;
-            objOld.NgayApDung = objUp.NgayApDung;   
-            objOld.NgayHetHan = objUp.NgayHetHan;
-            objOld.IdChiNhanh = objUp.IdChiNhanh;  
-            objOld.IdKhachHang = objUp.IdKhachHang;
-            objOld.IdNhanVien = objUp.IdNhanVien;  
-            objOld.IdPhong = objUp.IdViTriPhong;  
-            objOld.IdHoaDon = objUp.IdHoaDon;
-            objOld.TongTienHangChuaChietKhau = objUp.TongTienHangChuaChietKhau;
-            objOld.PTChietKhauHang = objUp.PTChietKhauHang;   
-            objOld.TongChietKhauHangHoa = objUp.TongChietKhauHangHoa;
-            objOld.TongTienHang = objUp.TongTienHang;
-            objOld.PTThueHD = objUp.PTThueHD;
-            objOld.TongTienThue = objUp.TongTienThue;
-            objOld.TongTienHDSauVAT = objUp.TongTienHDSauVAT;
-            objOld.PTGiamGiaHD = objUp.PTGiamGiaHD;
-            objOld.TongGiamGiaHD = objUp.TongGiamGiaHD;
-            objOld.ChiPhiTraHang = objUp.ChiPhiTraHang;
-            objOld.TongThanhToan = objUp.TongThanhToan;
-            objOld.ChiPhiHD = objUp.ChiPhiHD;
-            objOld.ChiPhi_GhiChu = objUp.ChiPhi_GhiChu;
-            objOld.DiemGiaoDich = objUp.DiemGiaoDich;
-            objOld.GhiChuHD = objUp.GhiChuHD;
-            objOld.TrangThai = objUp.TrangThai;
-            objOld.LastModifierUserId = AbpSession.UserId;
-            objOld.LastModificationTime = DateTime.Now;
-           
-            await _hoaDonRepository.UpdateAsync(objOld);
-            var dataHD = ObjectMapper.Map<CreateHoaDonDto>(objOld);
-            return dataHD;
-        } 
-        public async void Update_ChiTietHoaDon(List<HoaDonChiTietDto> lstCT)
+            return new CreateHoaDonDto();
+        }
+        /// <summary>
+        /// chỉ cập nhật chi tiết hóa đơn
+        /// </summary>
+        /// <param name="lstCT"></param>
+        /// <param name="idHoadon"></param>
+        /// <returns></returns>
+        public async Task Update_ChiTietHoaDon(List<HoaDonChiTietDto> lstCT, Guid idHoadon)
         {
+            var userID = AbpSession.UserId;
             List<BH_NhanVienThucHien> lstNVTH = new();
+
+            #region Delete ct if not exist
+            // comapre old & new
+            var lstOld = await _hoaDonChiTietRepository.GetAllListAsync(x => x.IdHoaDon == idHoadon);
+            var ctDelete = (from ctOld in lstOld
+                            join ctNew in lstCT on ctOld.Id equals ctNew.Id into delete
+                            from de in delete.DefaultIfEmpty()
+                            where de == null
+                            select ctOld.Id);
+            // update ct with TrangThai = 0 if delete
+            (await _hoaDonChiTietRepository.GetAllListAsync(x => ctDelete.Contains(x.Id))).ToList().ForEach(x =>
+            { x.TrangThai = 0; x.DeleterUserId = userID; x.DeletionTime = DateTime.Now; });
+            #endregion
+
             foreach (var item in lstCT)
             {
                 BH_HoaDon_ChiTiet ctUpdate = await _hoaDonChiTietRepository.FirstOrDefaultAsync(item.Id);
                 if (ctUpdate != null)
                 {
-                    ctUpdate.STT= item.STT;
-                    ctUpdate.SoLuong= item.STT;
-                    ctUpdate.IdDonViQuyDoi= item.IdDonViQuyDoi;
-                    ctUpdate.IdChiTietHoaDon= item.IdChiTietHoaDon;
-                    ctUpdate.DonGiaTruocCK= item.DonGiaTruocCK;
-                    ctUpdate.ThanhTienTruocCK= item.ThanhTienTruocCK;
-                    ctUpdate.PTChietKhau= item.PTChietKhau;
-                    ctUpdate.TienChietKhau= item.TienChietKhau;
-                    ctUpdate.DonGiaSauCK= item.DonGiaSauCK;
-                    ctUpdate.ThanhTienSauCK= item.ThanhTienSauCK;
-                    ctUpdate.PTThue= item.PTThue;
-                    ctUpdate.TienThue= item.TienThue;
-                    ctUpdate.DonGiaSauVAT= item.DonGiaSauVAT;
-                    ctUpdate.ThanhTienSauVAT= item.ThanhTienSauVAT;
-                    ctUpdate.GhiChu= item.GhiChu;
-                    ctUpdate.TrangThai= item.TrangThai;
-                    ctUpdate.LastModifierUserId = AbpSession.UserId;
+                    ctUpdate.STT = item.STT;
+                    ctUpdate.SoLuong = item.STT;
+                    ctUpdate.IdDonViQuyDoi = item.IdDonViQuyDoi;
+                    ctUpdate.IdChiTietHoaDon = item.IdChiTietHoaDon;
+                    ctUpdate.DonGiaTruocCK = item.DonGiaTruocCK;
+                    ctUpdate.ThanhTienTruocCK = item.ThanhTienTruocCK;
+                    ctUpdate.PTChietKhau = item.PTChietKhau;
+                    ctUpdate.TienChietKhau = item.TienChietKhau;
+                    ctUpdate.DonGiaSauCK = item.DonGiaSauCK;
+                    ctUpdate.ThanhTienSauCK = item.ThanhTienSauCK;
+                    ctUpdate.PTThue = item.PTThue;
+                    ctUpdate.TienThue = item.TienThue;
+                    ctUpdate.DonGiaSauVAT = item.DonGiaSauVAT;
+                    ctUpdate.ThanhTienSauVAT = item.ThanhTienSauVAT;
+                    ctUpdate.GhiChu = item.GhiChu;
+                    ctUpdate.TrangThai = item.TrangThai;
+                    ctUpdate.LastModifierUserId = userID;
                     ctUpdate.LastModificationTime = DateTime.Now;
                     await _hoaDonChiTietRepository.UpdateAsync(ctUpdate);
 
                     _nvthService.DeleteNVThucHienDichVu(item.Id);
 
-                    foreach (var nvth in item.nhanVienThucHien)
+                    if (item.nhanVienThucHien != null)
                     {
-                        BH_NhanVienThucHien nvNew = ObjectMapper.Map<BH_NhanVienThucHien>(nvth);
-                        nvNew.Id = Guid.NewGuid();
-                        nvNew.IdHoaDonChiTiet = ctUpdate.Id;
-                        nvNew.TenantId = AbpSession.TenantId ?? 1;
-                        nvNew.CreatorUserId = AbpSession.UserId;
-                        nvNew.CreationTime = DateTime.Now;
-                        lstNVTH.Add(nvNew);
+                        foreach (var nvth in item.nhanVienThucHien)
+                        {
+                            BH_NhanVienThucHien nvNew = ObjectMapper.Map<BH_NhanVienThucHien>(nvth);
+                            nvNew.Id = Guid.NewGuid();
+                            nvNew.IdHoaDonChiTiet = ctUpdate.Id;
+                            nvNew.TenantId = AbpSession.TenantId ?? 1;
+                            nvNew.CreatorUserId = userID;
+                            nvNew.CreationTime = DateTime.Now;
+                            lstNVTH.Add(nvNew);
+                        }
+                    }
+                }
+                else
+                {
+                    // insert new ct
+                    BH_HoaDon_ChiTiet ctNew = ObjectMapper.Map<BH_HoaDon_ChiTiet>(item);
+                    ctNew.Id = Guid.NewGuid();
+                    ctNew.IdHoaDon = idHoadon;
+                    ctNew.CreatorUserId = AbpSession.UserId;
+                    ctNew.CreationTime = DateTime.Now;
+                    await _hoaDonChiTietRepository.InsertAsync(ctNew);
+
+                    if (item.nhanVienThucHien != null)
+                    {
+                        foreach (var nvth in item.nhanVienThucHien)
+                        {
+                            BH_NhanVienThucHien nvNew = ObjectMapper.Map<BH_NhanVienThucHien>(nvth);
+                            nvNew.Id = Guid.NewGuid();
+                            nvNew.IdHoaDonChiTiet = ctNew.Id;
+                            nvNew.TenantId = AbpSession.TenantId ?? 1;
+                            nvNew.CreatorUserId = userID;
+                            nvNew.CreationTime = DateTime.Now;
+                            lstNVTH.Add(nvNew);
+                        }
                     }
                 }
             }
-            await _nvThucHien.InsertRangeAsync(lstNVTH);
+            if (lstNVTH.Count > 0)
+            {
+                await _nvThucHien.InsertRangeAsync(lstNVTH);
+            }
         }
         [HttpPost]
         public async Task DeleteHoaDon(Guid id)
@@ -298,6 +358,7 @@ namespace BanHangBeautify.HoaDon.HoaDon
                         item.IsDeleted = true;
                         item.DeleterUserId = AbpSession.UserId;
                         item.DeletionTime = DateTime.Now;
+                        item.TrangThai = 0;
                         await _hoaDonChiTietRepository.UpdateAsync(item);
                     }
                 }
@@ -316,18 +377,22 @@ namespace BanHangBeautify.HoaDon.HoaDon
                 hoaDon.IsDeleted = true;
                 hoaDon.DeleterUserId = AbpSession.UserId;
                 hoaDon.DeletionTime = DateTime.Now;
+                hoaDon.TrangThai = 0;
                 await _hoaDonRepository.UpdateAsync(hoaDon);
             }
         }
-        public async Task<List<PageHoaDonDto>> GetInforHoaDon_byId(Guid id) {
+        public async Task<List<PageHoaDonDto>> GetInforHoaDon_byId(Guid id)
+        {
             return await _repoHoaDon.GetInforHoaDon_byId(id);
         }
-        public async Task<List<PageHoaDonChiTietDto>> GetChiTietHoaDon_byIdHoaDon(Guid idHoaDon) {
+        public async Task<List<PageHoaDonChiTietDto>> GetChiTietHoaDon_byIdHoaDon(Guid idHoaDon)
+        {
             return await _repoHoaDon.GetChiTietHoaDon_byIdHoaDon(idHoaDon);
         }
 
         [HttpPost]
-        public async Task<PagedResultDto<PageHoaDonDto>> GetListHoaDon(HoaDonRequestDto param) {
+        public async Task<PagedResultDto<PageHoaDonDto>> GetListHoaDon(HoaDonRequestDto param)
+        {
             return await _repoHoaDon.GetListHoaDon(param, AbpSession.TenantId ?? 1);
         }
     }
